Test for custom mappings in batches.
